requirements.txt
typer==0.12.3
pyyaml==6.0.1
requests==2.31.0
tabulate==0.9.0
python-dateutil==2.8.2